dom/html/test/file_fullscreen-denied.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=545812

Test DOM fullscreen API.

-->
<head>
  <title>Test for Bug 545812</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript" src="file_fullscreen-utils.js"></script>
  <style>
  body {
    background-color: black;
  }
  </style>
</head>
<body>

<script type="application/javascript">

/** Test for Bug 545812 **/

function ok(condition, msg) {
  opener.ok(condition, "[denied] " + msg);
}

function is(a, b, msg) {
  opener.is(a, b, "[denied] " + msg);
}

function begin() {
  document.addEventListener("mozfullscreenchange", () => {
    ok(false, "Should never receive " +
       "a fullscreenchange event in the main window.");
  });
  SimpleTest.executeSoon(testIFrameWithoutAllowFullscreen);
}

function testIFrameWithoutAllowFullscreen() {
  window.continueTest = () => {
    delete window.continueTest;
    document.body.removeChild(iframe);
    // In the following tests we want to test trust context requirement
    // of fullscreen request, so temporary re-enable this pref.
    SpecialPowers.pushPrefEnv({
      "set":[["full-screen-api.allow-trusted-requests-only", true]]
    }, testNonTrustContext);
  };
  // Create an iframe without an allowfullscreen attribute, whose
  // contents request fullscreen. The request should be denied, and
  // we should not receive a fullscreenchange event in this document.
  var iframe = document.createElement("iframe");
  iframe.src = "file_fullscreen-denied-inner.html";
  document.body.appendChild(iframe);
}

function testNonTrustContext() {
  addFullscreenErrorContinuation(() => {
    ok(!document.mozFullScreen,
       "Should not grant request in non-trust context.");
    SimpleTest.executeSoon(testLongRunningEventHandler);
  });
  document.documentElement.mozRequestFullScreen();
}

function testLongRunningEventHandler() {
  function longRunningHandler() {
    window.removeEventListener("keypress", longRunningHandler);
    // Busy loop until 2s has passed. We should then be past the one
    // second threshold, and so our request for fullscreen should be
    // rejected.
    var end = (new Date()).getTime() + 2000;
    while ((new Date()).getTime() < end) {
      ; // Wait...
    }
    document.documentElement.mozRequestFullScreen();
  }
  addFullscreenErrorContinuation(() => {
    ok(!document.mozFullScreen,
       "Should not grant request in long-running event handler.");
    // Restore the pref environment we changed before
    // entering testNonTrustContext.
    SpecialPowers.popPrefEnv(finish);
  });
  window.addEventListener("keypress", longRunningHandler);
  synthesizeKey("VK_A", {});
}

function finish() {
  opener.nextTest();
}

</script>
</body>
</html>
